--- OneCoin/Server/src/main/java/OneCoin/Server/chat/chatRoom/entity/ChatRoom.java
@@ -1,13 +1,0 @@
-package OneCoin.Server.chat.chatRoom.entity;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-public class ChatRoom {
-    private Integer chatRoomId;
-    private Long numberOfChatters;
-}
