--- OneCoin/Server/src/main/java/OneCoin/Server/chat/chatRoom/entity/UserInChatRoom.java
@@ -1,15 +1,0 @@
-package OneCoin.Server.chat.chatRoom.entity;
-
-import lombok.*;
-
-import java.io.Serializable;
-
-@Builder
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserInChatRoom implements Serializable {
-    private String displayName;
-    private String email;
-}
