+++ OneCoin/Server/src/main/java/OneCoin/Server/chat/dto/ChatRequestDto.java
@@ -1,0 +1,19 @@
+package OneCoin.Server.chat.dto;
+
+import OneCoin.Server.chat.constant.MessageType;
+import lombok.*;
+
+import javax.validation.constraints.NotBlank;
+
+@Getter
+@Setter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@ToString
+public class ChatRequestDto {
+    @NotBlank
+    private Integer chatRoomId;
+    @NotBlank
+    private String message;
+}
