+++ OneCoin/Server/src/main/java/OneCoin/Server/chat/intercepter/MessageInterceptor.java
@@ -1,8 +1,9 @@
 package OneCoin.Server.chat.intercepter;
 
-import OneCoin.Server.chat.chatMessage.publisher.RedisPublisher;
-import OneCoin.Server.chat.chatRoom.service.ChatRoomService;
-import OneCoin.Server.chat.chatRoom.vo.UserInfoInChatRoom;
+import OneCoin.Server.chat.publisher.RedisPublisher;
+import OneCoin.Server.chat.service.ChatRoomService;
+import OneCoin.Server.chat.utils.AuthUtilForWebsocket;
+import OneCoin.Server.chat.vo.UserInfoInChatRoom;
 import OneCoin.Server.chat.constant.MessageType;
 import OneCoin.Server.config.auth.utils.LoggedInUserInfoUtilsForWebSocket;
 import OneCoin.Server.exception.BusinessLogicException;
