+++ OneCoin/Server/src/main/java/OneCoin/Server/chat/mapper/ChatRoomMapper.java
@@ -1,0 +1,10 @@
+package OneCoin.Server.chat.mapper;
+
+import OneCoin.Server.chat.entity.UserInChatRoom;
+import OneCoin.Server.user.entity.User;
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring")
+public interface ChatRoomMapper {
+    UserInChatRoom userToUserInChatRoom(User user);
+}
